Refactor CLI diagnostics and artifact IO modules
