+++ crates/perfgate-cli/src/git_context.rs
@@ -1,0 +1,93 @@
+//! Git and tracing context collected for repair diagnostics.
+
+use perfgate_types::{ChangedFilesSummary, OtelSpanIdentifiers, RepairGitMetadata};
+use std::collections::BTreeMap;
+use std::process::Command as ProcessCommand;
+
+pub(crate) fn otel_span_from_env() -> Option<OtelSpanIdentifiers> {
+    let trace_id = std::env::var("OTEL_TRACE_ID").ok();
+    let span_id = std::env::var("OTEL_SPAN_ID").ok();
+    if trace_id.is_none() && span_id.is_none() {
+        None
+    } else {
+        Some(OtelSpanIdentifiers { trace_id, span_id })
+    }
+}
+
+pub(crate) fn git_metadata() -> Option<RepairGitMetadata> {
+    let branch = run_git_capture(&["rev-parse", "--abbrev-ref", "HEAD"]);
+    let sha = run_git_capture(&["rev-parse", "HEAD"]);
+    if branch.is_none() && sha.is_none() {
+        None
+    } else {
+        Some(RepairGitMetadata { branch, sha })
+    }
+}
+
+pub(crate) fn changed_files_summary() -> Option<ChangedFilesSummary> {
+    let output = run_git_capture_bytes(&["status", "--porcelain", "-z"])?;
+    Some(parse_changed_files_summary(&output))
+}
+
+pub(crate) fn parse_changed_files_summary(output: &[u8]) -> ChangedFilesSummary {
+    let mut files = Vec::new();
+    let mut by_top = BTreeMap::new();
+
+    let mut entries = output
+        .split(|byte| *byte == b'\0')
+        .filter(|entry| !entry.is_empty());
+    while let Some(entry) = entries.next() {
+        if entry.len() <= 3 {
+            continue;
+        }
+
+        let status = &entry[..2];
+        let current_path = if status.iter().any(|code| matches!(code, b'R' | b'C')) {
+            entries.next().unwrap_or(&[])
+        } else {
+            &entry[3..]
+        };
+
+        if current_path.is_empty() {
+            continue;
+        }
+
+        let path = String::from_utf8_lossy(current_path).into_owned();
+        files.push(path.clone());
+        let top = path
+            .split(['/', '\\'])
+            .next()
+            .filter(|s| !s.is_empty())
+            .unwrap_or(".")
+            .to_string();
+        *by_top.entry(top).or_insert(0) += 1;
+    }
+
+    ChangedFilesSummary {
+        file_count: files.len() as u32,
+        files,
+        file_count_by_top_level: by_top,
+    }
+}
+
+pub(crate) fn run_git_capture(args: &[&str]) -> Option<String> {
+    let output = ProcessCommand::new("git").args(args).output().ok()?;
+    if !output.status.success() {
+        return None;
+    }
+    let text = String::from_utf8(output.stdout).ok()?;
+    let trimmed = text.trim().to_string();
+    if trimmed.is_empty() {
+        None
+    } else {
+        Some(trimmed)
+    }
+}
+
+fn run_git_capture_bytes(args: &[&str]) -> Option<Vec<u8>> {
+    let output = ProcessCommand::new("git").args(args).output().ok()?;
+    if !output.status.success() {
+        return None;
+    }
+    Some(output.stdout)
+}
